moved letterbox checking code into its own service
